Crate redstone

RedStone

redstone is a collection of utilities to make deserializing&decrypting RedStone payload. It includes a pure Rust implementation, along with extensions for certain networks.

Different crypto-mechanisms are easily injectable. The current implementation contains secp256k1- and k256-based variants.

Re-exports

pub use core::FeedValue;

Modules

contract

Contract specific logic.

core

default_ext

network

Structs

Bytes

Type wrapping bytes represantion.

FeedId

Type describing feed ids. We expect FeedId to be byte string like b“EUR“ converted to bytearray and padded with zeroes to the right.

RedStoneConfigImpl

SignerAddress

Type describing address of signer. Typically pubkey of length 20 bytes; As of right now we dont expect larger keys than 32 bytes. The address is normalized to contain only lowercase letters (A-F) -> (a-f).

TimestampMillis

Type describing timpestamp, we use to directly show we expect milliseconds.

Value

Enums

CryptoError

Traits

Avg

Crypto

RedStoneConfig

Configuration for the redstone protocol. Pluggable with custom environments and possible specialized crypto operations.