Struct redstone::core::config::Config

pub struct Config { /* private fields */ }

Configuration for a RedStone payload processor.

Specifies the parameters necessary for the verification and aggregation of values from various data points passed by the RedStone payload.

Implementations

impl Config

Auto-generated by derive_getters::Getters.

pub fn signer_count_threshold(&self) -> &u8

The minimum number of signers required validating the data.

Specifies how many unique signers (from different addresses) are required for the data to be considered valid and trustworthy.

pub fn signers(&self) -> &Vec<SignerAddress>

List of identifiers for signers authorized to sign the data.

Each signer is identified by a unique, network-specific byte string (Bytes), which represents their address.

pub fn feed_ids(&self) -> &Vec<FeedId>

Identifiers for the data feeds from which values are aggregated.

Each data feed id is represented by the FeedId type.

pub fn block_timestamp(&self) -> &TimestampMillis

The current block time in timestamp format, used for verifying data timeliness.

The value’s been expressed in milliseconds since the Unix epoch (January 1, 1970) and allows for determining whether the data is current in the context of blockchain time.

pub fn max_timestamp_delay_ms(&self) -> &TimestampMillis

The maximum delay of the package in regards to the current block in the blockchain.

The value’s been expressed in milliseconds since the Unix epoch (January 1, 1970).

pub fn max_timestamp_ahead_ms(&self) -> &TimestampMillis

The maximum time package was created ahead of the current block in the blockchain.

The value’s been expressed in milliseconds since the Unix epoch (January 1, 1970).

impl Config

pub fn try_new( signer_count_threshold: u8, signers: Vec<SignerAddress>, feed_ids: Vec<FeedId>, block_timestamp: TimestampMillis, max_timestamp_delay_ms: Option<TimestampMillis>, max_timestamp_ahead_ms: Option<TimestampMillis>, ) -> Result<Self, Error>

Verifies all members of the config.

This method checks whether all config members are correct.

Arguments

signer_count_threshold - The minimum number of signers required validating the data.

signers - List of identifiers for signers authorized to sign the data.

feed_ids - Identifiers for the data feeds from which values are aggregated.

block_timestamp - The current block time in timestamp format, used for verifying data timeliness.

max_timestamp_delay_ms - Maximum delay of the package against the current block timestamp. If None is provided then default config value is used.

max_timestamp_ahead_ms - Maximum ahead of time of the package against current block timestamp. If None is provided then default config value is used.

Returns

Success Self if arguments to the functions are correct or cresponding Err with redstone::network::Error otherwise.