redstone::core::processor_result
Struct ValidatedPayload

pub struct ValidatedPayload {
    pub timestamp: TimestampMillis,
    pub values: Vec<FeedValue>,
}

Represents the result of processing the RedStone payload.

This structure is used to encapsulate the outcome of a RedStone payload processing operation, particularly focusing on time-sensitive data and its associated values, according to the Config.

Fields
timestamp: TimestampMillis

The timestamp encountered during processing.

This field captures the time point (in milliseconds since the Unix epoch) among the processed data packages, indicating the starting boundary of the dataset’s time range.

values: Vec<FeedValue>

A collection of values processed during the operation.

Some feeds may be missing, if they did not meet some validation criteria.