Enum redstone::network::error::Error

pub enum Error {
    ContractError(Box<dyn ContractErrorContent>),
    NumberOverflow(U256),
    ArrayIsEmpty,
    CryptographicError(usize),
    SizeNotSupported(usize),
    WrongRedStoneMarker(Vec<u8>),
    NonEmptyPayloadRemainder(Vec<u8>),
    InsufficientSignerCount(usize, usize, U256),
    TimestampTooOld(usize, u64),
    TimestampTooFuture(usize, u64),
    ClonedContractError(u8, String),
}

Errors that can be encountered in the deserializing&decrypting the RedStone payload or just contract execution process.

These errors include issues with contract logic, data types, cryptographic operations, and conditions specific to the requirements.

Variants

ContractError(Box<dyn ContractErrorContent>)

Represents errors that arise from the contract itself.

This variant is used for encapsulating errors that are specific to the contractโs logic or execution conditions that arenโt covered by more specific error types.

NumberOverflow(U256)

Indicates an overflow error with U256 numbers.

Used when operations on U256 numbers exceed their maximum value, potentially leading to incorrect calculations or state.

ArrayIsEmpty

Used when an expected non-empty array or vector is found to be empty.

This could occur in scenarios where the contract logic requires a non-empty collection of items for the correct operation, for example, during aggregating the values.

CryptographicError(usize)

Represents errors related to cryptographic operations.

This includes failures in signature verification, hashing, or other cryptographic processes, with the usize indicating the position or identifier of the failed operation.

SizeNotSupported(usize)

Signifies that an unsupported size was encountered.

This could be used when a data structure or input does not meet the expected size requirements for processing.

WrongRedStoneMarker(Vec<u8>)

Indicates that the marker bytes for RedStone are incorrect.

This error is specific to scenarios where marker or identifier bytes do not match expected values, potentially indicating corrupted or tampered data.

NonEmptyPayloadRemainder(Vec<u8>)

Used when there is leftover data in a payload that should have been empty.

This could indicate an error in data parsing or that additional, unexpected data was included in a message or transaction.

InsufficientSignerCount(usize, usize, U256)

Indicates that the number of signers does not meet the required threshold.

This variant includes the current number of signers, the required threshold, and potentially a feed_id related to the operation that failed due to insufficient signers.

TimestampTooOld(usize, u64)

Used when a timestamp is older than allowed by the processor logic.

Includes the position or identifier of the timestamp and the threshold value, indicating that the provided timestamp is too far in the past.

TimestampTooFuture(usize, u64)

Indicates that a timestamp is further in the future than allowed.

Similar to TimestampTooOld, but for future timestamps exceeding the contractโs acceptance window.

ClonedContractError(u8, String)

Represents errors that need to clone ContractErrorContent, which is not supported by default.

This variant allows for the manual duplication of contract error information, including an error code and a descriptive message.

Implementations

impl Error

pub fn contract_error<T: ContractErrorContent + 'static>(value: T) -> Error

pub(crate) fn code(&self) -> u16