Enum redstone::network::error::Error

pub enum Error {

    ContractError(ContractErrorContent),
    NumberOverflow(Value),
    ArrayIsEmpty,
    CryptographicError(CryptoError),
    SizeNotSupported(usize),
    WrongRedStoneMarker(Vec<u8>),
    NonEmptyPayloadRemainder(usize),
    InsufficientSignerCount(usize, usize, FeedId),
    TimestampTooOld(usize, TimestampMillis),
    TimestampTooFuture(usize, TimestampMillis),
    ReocuringFeedId(FeedId),
    ConfigInsufficientSignerCount(u8, u8),
    ConfigExceededSignerCount(usize, usize),
    ConfigReocuringSigner(SignerAddress),
    ConfigEmptyFeedIds,
    ConfigReocuringFeedId(FeedId),
    TimestampMustBeGreaterThanBefore,
    CurrentTimestampMustBeGreaterThanLatestUpdateTimestamp,

}

Errors that can be encountered in the deserializing&decrypting the RedStone payload or just contract execution process.

These errors include issues with contract logic, data types, cryptographic operations, and conditions specific to the requirements.

Variants

ContractError(ContractErrorContent)

Represents errors that arise from the contract itself.

This variant is used for encapsulating errors that are specific to the contractโs logic or execution conditions that arenโt covered by more specific error types.

NumberOverflow(Value)

Indicates an overflow error with U256 numbers.

Used when operations on U256 numbers exceed their maximum value, potentially leading to incorrect calculations or state.

ArrayIsEmpty

Used when an expected non-empty array or vector is found to be empty.

This could occur in scenarios where the contract logic requires a non-empty collection of items for the correct operation, for example, during aggregating the values.

CryptographicError(CryptoError)

Represents errors related to cryptographic operations.

This includes failures in signature verification, hashing, or other cryptographic processes.

SizeNotSupported(usize)

Signifies that an unsupported size was encountered.

This could be used when a data structure or input does not meet the expected size requirements for processing.

WrongRedStoneMarker(Vec<u8>)

Indicates that the marker bytes for RedStone are incorrect.

This error is specific to scenarios where marker or identifier bytes do not match expected values, potentially indicating corrupted or tampered data.

NonEmptyPayloadRemainder(usize)

Used when there is leftover data in a payload that should have been empty.

This could indicate an error in data parsing or that additional, unexpected data was included in a message or transaction.

InsufficientSignerCount(usize, usize, FeedId)

Indicates that the number of signers does not meet the required threshold.

This variant includes the current number of signers, the required threshold, and potentially a feed_id related to the operation that failed due to insufficient signers.

TimestampTooOld(usize, TimestampMillis)

Used when a timestamp is older than allowed by the processor logic.

Includes the position or identifier of the timestamp and the threshold value, indicating that the provided timestamp is too far in the past.

TimestampTooFuture(usize, TimestampMillis)

Indicates that a timestamp is further in the future than allowed.

Similar to TimestampTooOld, but for future timestamps exceeding the contractโs acceptance window.

ReocuringFeedId(FeedId)

Indicates that a FeedId is reocuring in data points.

Includes FeedId that is reocuring.

ConfigInsufficientSignerCount(u8, u8)

ConfigInsufficientSignerCount occurs when number of signer count is not at least equal required signer threshold.

Includes current config signer list length and minimum required signer count.

ConfigExceededSignerCount(usize, usize)

ConfigExceededSignerCount occurs when number of signer count is larger than config max allowed signer count. Look in to core::config crate to acknowladge constant value.

Includes current config signer list length and maximum allowed signer count per config.

ConfigReocuringSigner(SignerAddress)

Indicates that a SignerAddress is reocuring on config signer list.

Includes SignerAddress that is reocuring.

ConfigEmptyFeedIds

Indicates that list doesnโt contain FeedIds.

ConfigReocuringFeedId(FeedId)

Indicates that a FeedId is reocuring on config feed_ids list.

Includes FeedId that is reocuring.

TimestampMustBeGreaterThanBefore

Indicates that the provided timestamp is not greater than a previously written timestamp.

For the price adapter to accept a new price update, the associated timestamp must be strictly greater than the timestamp of the last update. This error is raised if a new timestamp does not meet this criterion, ensuring the chronological integrity of price data.

CurrentTimestampMustBeGreaterThanLatestUpdateTimestamp

Indicates that the current timestamp is not greater than the timestamp of the last update.

This error is raised to ensure that the data being written has a timestamp strictly greater than the most recent timestamp already stored in the system. It guarantees that new data is not outdated or stale compared to the existing records, thereby maintaining the chronological integrity and consistency of the updates.

Implementations

impl Error

pub fn code(&self) -> u16